+++ app/src/main/java/com/bll/lnkstudy/mvp/model/paper/ExamScoreItem.java
@@ -8,6 +8,6 @@ public class ExamScoreItem implements Serializable {
     public String score="0";
     public int sort;
     public int result;//0错1对
-    public int label;//题目标准分数
+    public double label;//题目标准分数
     public List<ExamScoreItem> childScores=new ArrayList<>();
 }
